NLP.RAKE.Text

Description

The RAKE Text interface

Keywords

type WordScore = (Text, Double)

The result is a keyword candidate, a keyword consisting of one or more words and a score associated with this keyword.

candidates :: StopwordsMap -> String -> [Text] -> [WordScore]

This interface provides most flexibility. It expects a Map of stop words, a nosplit list used by the word splitter and a text split into phrases. Users may pass in their own stop word list (e.g. by loading it from a file, see loadStopWords) or one of the predefined lists (smartStopwords, foxStopwords).

keywords :: Text -> [WordScore]

The keywords function is a convenience interface that takes a couple of decisions internally: it uses the defaultStoplist and the English language nosplit list and it splits the text into phrases using the pSplitter.

The function is equivalent to

 candidates defaultStoplist defaultNosplit . pSplitter

rawKeywords :: [WordScore] -> [Text]

Given a WordScore list rawKeywords yields a list of the keywords without the score (it, hence, maps fst to the WordScore list).

Utitlities

sortByScore :: [WordScore] -> [WordScore]

Sort the WordScore list by scores (descending!)

sortByWord :: [WordScore] -> [WordScore]

Sort the WordScore list by words (ascending!)

pSplitter :: Text -> [Text]

Default phrase splitter. It splits phrases at characters in the punctuation category (those for which isPunctuation is True) with the exception of '-'.

defaultNosplit :: String

List containing Char at which we do not split words. Note that this list is based on English.

Stopwords

The very heart of the RAKE algorithm is the use of stop words, a concept defined by NLP pioneer Hans Peter Luhn. Stop words are frequent words in a language that are considered to be void of specific semantics. They, of course, have an important role in the language, but they do not identify the topic a specific document is about, e.g. "is", "the", "of" and so on. Stop words depend on the specific context of the documents to be analysed; there are, however, frequently used lists with wide applicability.

The library comes with two stop word lists built in: the smartStoplist and the foxStoplist, both for English. The list used by default is smartStoplist.

The user is free to define her own stop word list, which can be loaded from a file using loadStopWords. The file format is simple:

Lines starting with '#' are ignored (comments);

Each line contains one word.

type StopwordsMap = Map Text ()

Search tree for stop words

mkStopwords :: [Text] -> StopwordsMap

Make StopwordsMap starting from a list of stop words encoded as Text

mkStopwordsStr :: [String] -> StopwordsMap

Make StopwordsMap starting from a list of stop words encoded as String

stopword :: StopwordsMap -> Text -> Bool

Search for a chunk of Text in the StopwordsMap. Note that there is a list of symbols, the nolist, that still count as stop words (e.g. "-").

defaultStoplist :: StopwordsMap

The default stop word list (smartStoplist).

smartStoplist :: StopwordsMap

The "smart" stop word list

foxStoplist :: StopwordsMap

The "Fox" stop word list

loadStopWords :: FilePath -> IO StopwordsMap

Load a stop word list from a file.